added some improvements to api tests

+++ packages/api/src/utlis/jest.ts
@@ -7,6 +7,10 @@ import { getLocals } from './ILocals';
 export const jestCleanUp = async ( app: Express ) => {
   const { dbClient, MongoDB } = getLocals( app );
 
+  if ( MongoDB.databaseName.indexOf( '-dHdjE4FMoP-' ) === -1 ) {
+    throw new Error( 'QeKIhB_q3s | db name must have "-dHdjE4FMoP-" marking present' );
+  }
+
   await MongoDB.dropDatabase();
   await dbClient.close();
 };
